// src/types/GoalProgress.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Goal } from "./Goal";
import type { MilestoneProgress } from "./MilestoneProgress";

export type GoalProgress = { goal: Goal, milestones: Array<MilestoneProgress>, done: number, total: number, };